fix(dashboard): split 'use server' file — move types/helpers to dashboard.types.ts
Next.js requires every export in a 'use server' file to be async.
Non-async exports (types, consts, computeFreshnessStatus, secondsSince)
have been moved to a sibling dashboard.types.ts with no directive.
actions.ts re-exports all types for backwards-compatible imports.

Files changed:
- NEW  apps/frontend/src/app/dashboard/dashboard.types.ts
- EDIT apps/frontend/src/app/dashboard/actions.ts (types stripped, re-exported)
- EDIT apps/frontend/src/app/dashboard/actions.test.ts (import source updated)
- EDIT src/components/Dashboard/DashboardFreshnessBadge.tsx (import updated)
- EDIT src/components/Dashboard/__tests__/DashboardFreshnessBadge.test.tsx (import updated)

All 317 tests pass.
